adding import from csv and sample chartjs

+++ src/prj/myapp/models/module.py
@@ -5,9 +5,9 @@ from django.db import models
 class Module(models.Model):
     code = models.CharField(max_length=5, unique=True)
     name = models.CharField(max_length=256)
-    cm_hours = models.IntegerField()
-    td_hours = models.IntegerField()
-    tp_hours = models.IntegerField()
+    cm_hours = models.FloatField()
+    td_hours = models.FloatField()
+    tp_hours = models.FloatField()
 
     def __str__(self):
         return str(self.code) + ' --- ' + str(self.name)
